service/collect-pi-data/requirements.txt
psycopg2-binary
scholarly
biopython